% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_distribution.R
\name{beta_interval}
\alias{beta_interval}
\title{Apply a confidence interval to values in a dataframe based on the beta distribution.}
\usage{
beta_interval(.data, alpha, beta, conf = 0.95)
}
\arguments{
\item{.data}{Data frame; a confidence interval will be applied to each row in
the data frame.}

\item{alpha}{First shape parameter of the beta distribution, must be greater than 0}

\item{beta}{Second shape parameter of the beta distribution, must be greater than 0}

\item{conf}{Confidence interval, must be between \verb{[0, 1]}. Defaults to 0.95.}
}
\description{
Apply a confidence interval to values in a dataframe based on the beta distribution.
}
\examples{
\dontrun{
# create example df
alpha <- rnorm(10, 100, 20)
beta <- rnorm(10, 80, 10)
example <- dplyr::bind_cols(alpha, beta)
colnames(example) <- c("alpha", "beta")

# apply the default confidence interval of 0.95
beta_interval(example, alpha, beta)

# apply a custom confidence interval
beta_interval(example, alpha, beta, conf = 0.99)

}

}
