% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_distribution.R
\name{beta_lower}
\alias{beta_lower}
\title{Util function for finding the lower bound of a confidence interval on a beta distribution}
\usage{
beta_lower(alpha, beta, conf)
}
\arguments{
\item{alpha}{First shape parameter of the beta distribution, must be greater than 0}

\item{beta}{Second shape parameter of the beta distribution, must be greater than 0}

\item{conf}{Confidence interval, must be between \verb{[0, 1]}.}
}
\description{
Util function for finding the lower bound of a confidence interval on a beta distribution
}
