% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{cholesky_decompose}
\alias{cholesky_decompose}
\title{The Cholesky Decomposition}
\usage{
cholesky_decompose(x, ...)
}
\arguments{
\item{x}{A symmetric, positive-definite square matrix.}

\item{...}{Additional parameters to pass to \code{chol()}.}
}
\value{
A lower-triangular square matrix the same size as \code{x}.
}
\description{
Compute the Cholesky factorization of a real, symmetric,
positive-definite square matrix. Returns the lower-left triangular matrix
to match Stan's version. This function is simply an opinionated wrapper of
the base R function, \code{chol()}.
}
\examples{
x <- matrix(c(2, 1, 1, 2), nrow = 2)
cholesky_decompose(x)
}
