% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse_extensions.R
\name{color_text}
\alias{color_text}
\title{Color text using HTML color styling}
\usage{
color_text(text, hex_color_code)
}
\arguments{
\item{text}{text to be colored}

\item{hex_color_code}{color to be applied to the text}
}
\description{
Useful in conjuction with ggtext and glue
}
\examples{
color_text("Hello there!", "#BD43BF")
}
