% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{condition_gaussian_process}
\alias{condition_gaussian_process}
\title{Condition a Gaussian Process}
\usage{
condition_gaussian_process(
  n,
  x,
  y,
  x_new,
  amplitude = 1,
  length_scale = 1,
  delta = 1e-09
)
}
\arguments{
\item{n}{number of random draws.}

\item{x}{position of values to condition on.}

\item{y}{values to condition on (at positions \code{x}).}

\item{x_new}{new positions to estimate values at.}

\item{amplitude}{Numeric, controls the vertical scale of the covariance function}

\item{length_scale}{Numeric, controls the horizontal scale of the covariance function}

\item{delta}{A small offset added along the diagonal to ensure that the
function returns a positive-semidefinite matrix. Larger values allow for
increased variation at individual positions along the vector \code{x}.}
}
\value{
A matrix of size \eqn{M \times N}, where \eqn{M} is \code{length(x_new)},
and \eqn{N} is \code{n}. Each column is a random draw from the conditioned
multivariate normal where each row corresponds to the draw at the position
in \code{x_new}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Condition a Multivariate Normal distribution based on a
Gaussian Process using a exponentiated quadratic covariance matrix.
}
\examples{
x <- 1:10
y <- rep(0, 10)
x_new <- 11:15

condition_gaussian_process(5, x, y, x_new)
}
