% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_extensions.R
\name{cramer_v}
\alias{cramer_v}
\title{Calculate the effect size for a Chi-Square test using the Cramer's V measure}
\usage{
cramer_v(chi_sq, n, deg_free)
}
\arguments{
\item{chi_sq}{Chi-Square statistic from a Chi-Square test}

\item{n}{total number of observations}

\item{deg_free}{degrees of freedom in the Chi-Square test}
}
\description{
Calculate the effect size for a Chi-Square test using the Cramer's V measure
}
\examples{
\dontrun{
cramer_v(33.2, 125, 4)
}
}
