% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_extensions.R
\name{deduplicate_ip}
\alias{deduplicate_ip}
\title{Remove duplicate responses from the Inpatient survey}
\usage{
deduplicate_ip(.data)
}
\arguments{
\item{.data}{A tibble returned by \code{fetch_surveys()}. The column \code{survey_id}
may be removed but both \code{survey_name} and \code{responses} must be present.}
}
\description{
Due to Qualtrics' HCAHPs sampling engine, Memorial Hermann's Inpatient survey
is a master file containing responses for Inpatient, Pedi Inpatient, and
Inpatient Rehab (Inpatient Rehab was separated out to its own standing survey
in early 2022, but historical responses are duplicated across both the
Inpatient and Inpatient Rehab survey files). \code{deduplicate_ip()} takes in a
tibble from \code{fetch_surveys()} and removes duplicated responses from the
Inpatient file, using the encounter ID (\code{"UNIQUE_ID"}) as the key.
}
\examples{
\dontrun{
# pull in surveys from Qualtrics
surveys <-
  c("Live - Inpatient",
    "Live - Pediatric Inpatient",
    "Live - Inpatient Rehab")

surveys <- fetch_surveys(surveys)

# deduplicate records
deduplicate_ip(surveys)
}
}
