% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_transforms.R
\name{expit}
\alias{expit}
\title{Transform a value on an unbounded scale to a bound scale}
\usage{
expit(x, lower_bound = 0, upper_bound = 1, base = exp(1))
}
\arguments{
\item{x}{Unbounded value to be transformed}

\item{lower_bound}{Lower bound of new scale; defaults to 0.}

\item{upper_bound}{Upper bound of new scale; defaults to 1.}

\item{base}{Base used for transformation; defaults to exp(1) = e}
}
\description{
Transform a value on an unbounded scale to a bound scale
}
\details{
Transform a value on an unbounded scale (i.e., \verb{[-Inf, Inf]}) to a
bound scale. Passing \code{-Inf}/\code{Inf} will return the lower/upper bounds,
respectively. Uses the natural base, \code{e}, by default, but other bases can be
passed to the parameter \code{base}
}
\examples{
\dontrun{
# bound vector of unbounded values to [0, 1]
myvec <- rnorm(10)
expit(myvec)

# bound vector of unbounded values to [100, 200]
expit(myvec, lower_bound = 100, upper_boun = 200)
}
}
