% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_extensions.R
\name{fetch_survey_named}
\alias{fetch_survey_named}
\title{Fetch a Qualtrics survey by name via the Qualtrics API}
\usage{
fetch_survey_named(survey_name, ...)
}
\arguments{
\item{survey_name}{Survey name as it appears in the Qualtrics UI}

\item{...}{Additional arguments to be passed to \code{qualtRics::fetch_survey()}}
}
\description{
Fetch a Qualtrics survey by name via the Qualtrics API
}
\examples{
\dontrun{
# fetch all survey responses
mysurvey <- fetch_survey_named("my survey name")

# fetch survey based on return date
mysurvey <- fetch_survey_named("my survey name", start_date = "2021-07-01", end_date = "2021-07-30")
}

}
