% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_extensions.R
\name{fetch_surveys}
\alias{fetch_surveys}
\title{Fetch a list of Qualtrics surveys by name via the Qualtrics API}
\usage{
fetch_surveys(survey_names, ...)
}
\arguments{
\item{survey_names}{A vector of survey names as they appear in the Qualtrics UI.}

\item{...}{Additional arguments to be passed to \code{qualtRics::fetch_survey()}.}
}
\value{
For a single value supplied to \code{survey_names}: a tibble of the survey
responses from \code{qualtRics::fetch_survey()}. For multiple \code{survey_names}, a
tibble with colums \code{survey_name}, \code{survey_id}, and \code{responses}. The \code{responses}
contains the responses for each survey as returned by
\code{qualtRics::fetch_survey()}.
}
\description{
Fetch a list of Qualtrics surveys by name via the Qualtrics API
}
\examples{
\dontrun{
# fetch a single survey by name
my_survey <- fetch_surveys("my cool survey")

# fetch multiple surveys by name
my_surveys <- fetch_surveys(c("my cool survey", "my lame survey"))
}
}
