% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_extensions.R
\name{fix_survey_names}
\alias{fix_survey_names}
\title{Reformat Memorial Hermann Survey Names}
\usage{
fix_survey_names(.data, survey_names)
}
\arguments{
\item{.data}{A tibble or dataframe containing survey names.}

\item{survey_names}{The name of the column in \code{.data} containing survey names.}
}
\description{
Convenience function for reformatting MH Survey names as they appear in the
Qualtrics UI to something more legible.
}
\examples{
# list of survey names as they appear in the qualtrics ui
surveys <-
  tibble::tibble(
    names = c("Live - Adult Day Surgery",
              "Live - Adult Emergency Department",
              "Live - Adult Inpatient Rehab",
              "Live - Adult Medical Practice",
              "Live - Adult Outpatient",
              "Live - Adult Outpatient - Oncology",
              "Live - Adult Outpatient - Vaccine",
              "Live - Adult Outpatient Rehab",
              "Live - Adult Telemedicine",
              "Live - Adult Urgent Care",
              "Live - HCAHPS - Paper",
              "Live - Home Medical Equipment",
              "Live - Infusion Pharmacy",
              "Live - Inpatient",
              "Live - Pediatric - Emergency Department",
              "Live - Pediatric Inpatient")
  )

# bad names!
surveys

# better names!
fix_survey_names(surveys, names)
}
