% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_transforms.R
\name{logit}
\alias{logit}
\alias{expit}
\title{Transform a between bound and unbounded scales.}
\usage{
logit(x, lower_bound = 0, upper_bound = 1, base = exp(1))

expit(x, lower_bound = 0, upper_bound = 1, base = exp(1))
}
\arguments{
\item{x}{Value to be rescaled.}

\item{lower_bound}{Lower bound of x; defaults to 0.}

\item{upper_bound}{Upper bound of x; defaults to 1.}

\item{base}{Base used for log-transform, defaults to \code{exp(1)} = e}
}
\description{
Transform a between bound and unbounded scales.
}
\details{
\code{logit()} will transform a value on a bound scale to an unbound scale,
while \code{expit()} performs the reverse transformation. Values at the lower/upper
bounds correspond to \code{-Inf}/\code{Inf} on the unbounded scale. Uses the natural base,
\code{e}, by default, but other bases can be passed to the parameter \code{base}.
}
\examples{
# unbound vector of values bound between [0, 1]
myvec <- seq(0, 10, 1)/10
logit(myvec)

# unbound vector of values between [100, 200]
myvec <- seq(100, 200, 10)
logit(myvec, lower_bound = 100, upper_bound = 200)

# bound vector of unbounded values to [0, 1]
myvec <- rnorm(10)
expit(myvec)

# bound vector of unbounded values to [100, 200]
expit(myvec, lower_bound = 100, upper_bound = 200)
}
