% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse_extensions.R
\name{ncells}
\alias{ncells}
\title{Determine the number of \code{cells} in a tibble or dataframe}
\usage{
ncells(.data)
}
\arguments{
\item{.data}{a tibble or dataframe}
}
\description{
Determine the number of \code{cells} in a tibble or dataframe
}
\examples{
# return the number of cells in the `iris` dataset
cells(iris)
}
