% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_extensions.R
\name{normal_interval}
\alias{normal_interval}
\title{Apply a confidence interval to values in a dataframe based on the normal distribution.}
\usage{
normal_interval(.data, mean, std_dev, conf = 0.95)
}
\arguments{
\item{.data}{Data frame; a confidence interval will be applied to each row in
the data frame.}

\item{mean}{Mean of the distribution}

\item{std_dev}{Standard deviation of the distribution}

\item{conf}{Confidence interval, must be between \verb{[0, 1]}. Defaults to 0.95.}
}
\description{
Apply a confidence interval to values in a dataframe based on the normal distribution.
}
\examples{
\dontrun{
# create example df
mean <- rnorm(10, 100, 20)
std_dev <- rnorm(10, 80, 10)
example <- dplyr::bind_cols(mean, std_dev)
colnames(example) <- c("mean", "std_dev")

# apply the default confidence interval of 0.95
normal_interval(example, mean, std_dev)

# apply a custom confidence interval
normal_interval(example, mean, std_dev, conf = 0.99)

}

}
