% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_extensions.R
\name{normal_upper}
\alias{normal_upper}
\title{Util function for finding the upper bound of a confidence interval on a normal distribution}
\usage{
normal_upper(mean, std_dev, conf)
}
\arguments{
\item{mean}{Mean of the distribution}

\item{std_dev}{Standard deviation of the distribution}

\item{conf}{Confidence interval, must be between \verb{[0, 1]}}
}
\description{
Util function for finding the upper bound of a confidence interval on a normal distribution
}
