% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_extensions.R
\name{pct_diff}
\alias{pct_diff}
\title{Calculate the percent difference between two values}
\usage{
pct_diff(x, y)
}
\arguments{
\item{x, y}{Values you want to determine the percent difference between.}
}
\description{
Calculate the percent difference between two values
}
\examples{
pct_diff(182, 195)
}
