% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse_extensions.R
\name{percent}
\alias{percent}
\title{Return the percentage each value or combination of values appear in a tibble or dataframe}
\usage{
percent(.data, ..., .keep_n = FALSE)
}
\arguments{
\item{.data}{a tibble or dataframe}

\item{...}{parameters to pass to \code{dplyr::count()}}

\item{.keep_n}{logical; keep or remove the \code{n} col created by \code{dplyr::count()}. Defaults to \code{FALSE}.}
}
\description{
Return the percentage each value or combination of values appear in a tibble or dataframe
}
\examples{
\dontrun{
# return a tibble of the percentage of gear type in the `mtcars` dataset
percent(mtcars, gear)
}
}
