% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workboots.R
\name{predict_boot_internal}
\alias{predict_boot_internal}
\title{Fit a model and predict based on a single bootstrap resample}
\usage{
predict_boot_internal(workflow, boot_splits, new_data, index, seed = 123)
}
\arguments{
\item{workflow}{An un-fitted workflow object.}

\item{boot_splits}{A bootstrap split object created by \code{rsample::bootstraps()}.}

\item{new_data}{New data to make predictions}

\item{index}{Index of \code{boot_splits} to use for training}

\item{seed}{A single seed value to be passed to \code{set.seed()}.}
}
\description{
Fit a model and predict based on a single bootstrap resample
}
