% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workboots.R
\name{predict_boots}
\alias{predict_boots}
\title{Fit and predict a workflow using many bootstrap resamples.}
\usage{
predict_boots(
  workflow,
  training_data,
  new_data,
  n_models = 100,
  seed = 123,
  ...
)
}
\arguments{
\item{workflow}{An un-fitted workflow object.}

\item{training_data}{A tibble or dataframe of data to be resampled and used for training.}

\item{new_data}{A tibble or dataframe used to make predictions.}

\item{n_models}{An integer for the number of bootstrap resampled models that will be created.}

\item{seed}{A single seed value to be passed to \code{set.seed()}.}

\item{...}{Additional params passed to \code{rsample::bootstraps()}.}
}
\description{
Fit and predict a workflow using many bootstrap resamples.
}
\examples{
\dontrun{
library(tidymodels)

# setup a workflow without fitting
wf <-
  workflow() \%>\%
  add_recipe(recipe(qsec ~ wt, data = mtcars)) \%>\%
  add_model(linear_reg())

# fit and predict 125 bootstrap resampled models to mtcars
wf \%>\%
  predict_boots(training_data = mtcars, new_data = mtcars, n_models = 125)
}
}
