% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riekelib-package.R
\docType{package}
\name{riekelib-package}
\alias{riekelib}
\alias{riekelib-package}
\title{riekelib: A collection of personal functions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of functions for personal/professional use, not super well documented or tested. Depending on how the package grows, some functions (or groups of functions) may be formalized in their own succinct packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/markjrieke/riekelib}
  \item \url{https://markjrieke.github.io/riekelib/}
  \item Report bugs at \url{https://github.com/markjrieke/riekelib/issues}
}

}
\author{
\strong{Maintainer}: Mark Rieke \email{markjrieke@gmail.com}

}
\keyword{internal}
