% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps.R
\name{rnps}
\alias{rnps}
\title{Random generation of NPS based on counts of promoters, passives, and detractors}
\usage{
rnps(n, promoters, passives, detractors)
}
\arguments{
\item{n}{number of observations to generate}

\item{promoters, passives, detractors}{counts of promoters, passives, and detractors}
}
\description{
Random generation of NPS based on counts of promoters, passives, and detractors
}
\examples{
\dontrun{
# generate 10000 nps scores based on a sample of 100 respondents
rnps(10000, 70, 20, 10)
}
}
