% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_extensions.R
\name{scale_xy_percent}
\alias{scale_xy_percent}
\alias{scale_x_percent}
\alias{scale_y_percent}
\alias{scale_xy_comma}
\alias{scale_x_comma}
\alias{scale_y_comma}
\title{Scale shortcuts}
\usage{
scale_xy_percent(..., accuracy = 1)

scale_x_percent(..., accuracy = 1)

scale_y_percent(..., accuracy = 1)

scale_xy_comma(..., log10 = FALSE)

scale_x_comma(..., log10 = FALSE)

scale_y_comma(..., log10 = FALSE)
}
\arguments{
\item{...}{parameters passed to \verb{scale_*_continuous}}

\item{accuracy}{passed to \code{scales::label_percent()}}

\item{log10}{should the axes be on the log10 scale? Defaults to \code{FALSE}}
}
\description{
Format the labels on the x-axis, y-axis, or both
}
\examples{
library(ggplot2)

# quickly add comma format to x/y scales
tibble::tibble(x = rlnorm(1e3, 10),
               y = rlnorm(1e3, 10)) |>
  ggplot(aes(x, y)) +
  geom_point() +
  scale_xy_comma(log10 = TRUE)

# quickly add percent format to x/y scales
tibble::tibble(x = rbeta(1e3, 2, 2),
               y = rbeta(1e3, 2, 2)) |>
  ggplot(aes(x, y)) +
  geom_point() +
  scale_xy_percent()
}
