% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_extensions.R
\name{set_recode_path}
\alias{set_recode_path}
\title{Set a path for retrieving MH recodes}
\usage{
set_recode_path(path = NULL, install = TRUE)
}
\arguments{
\item{path}{folder path to recode list. Should include the full file name, i.e.,
\code{"path/to/recode/list/list_file.csv"}.}

\item{install}{save the folder path to your .Renviron for future use.}
}
\description{
Saves an environment variable containing the path to the recode list. Saved as
\code{MEMORIAL_HERMANN_RECODE_PATH}.
}
\examples{
\dontrun{
  set_recode_path("made/up/path/file.csv")
}
}
