% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_transforms.R
\name{softmax}
\alias{softmax}
\title{Apply the softmax function to a vector of values.}
\usage{
softmax(x)
}
\arguments{
\item{x}{Vector of values.}
}
\description{
Apply the softmax function to a vector of values.
}
\details{
Apply the softmax function to a vector of values, e.g.:

\deqn{\text{softmax}(z_i) = \frac{e^{z_i}}{\sum_{j = 1}^K \ e^{z_j}}}
}
\examples{
softmax(seq(from = 0, to = 1, by = 0.25))
}
