% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_extensions.R
\name{standardize}
\alias{standardize}
\title{Center and scale a set of values.}
\usage{
standardize(x)
}
\arguments{
\item{x}{A vector of values to be scaled.}
}
\description{
Transform values in \code{x} such that the mean is approximately 0 and the standard
deviation is approximately 1. Note that this is a whole-cloth reimplemntation
of a function of the same name from the \href{https://github.com/rmcelreath/rethinking/tree/master}{rethinking}
package by \href{https://xcelab.net/rm/}{Dr. Richard McElreath}.
}
\examples{
x <- rnorm(1000, mean = 3, sd = 4)
z <- standardize(x)

mean(z)
sd(z)
}
