% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sys_sleep}
\alias{sys_sleep}
\title{Sleep with a progress bar}
\usage{
sys_sleep(time, ..., message = "Sleeping...", .progress = TRUE)
}
\arguments{
\item{time}{the amount of time (in seconds) to sleep}

\item{...}{additional parameters passed to \code{cli::cli_progress_bar()}}

\item{message}{a message to display alongside the progress bar}

\item{.progress}{whether or not to sleep with a progress bar. Defaults to \code{TRUE}.}
}
\description{
A small wrapper around \code{Sys.sleep()} that adds a progress bar.
}
\examples{
# sleep with a progress bar
sys_sleep(3)
}
