% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workboots.R
\name{vi_boots}
\alias{vi_boots}
\title{Fit and estimate variable importance from a workflow using many bootstrap resamples.}
\usage{
vi_boots(workflow, training_data, n_models = 100, seed = 123, ...)
}
\arguments{
\item{workflow}{An un-fitted workflow object.}

\item{training_data}{A tibble or dataframe of data to be resampled and used for training.}

\item{n_models}{An integer for the number of bootstrap resampled models that will be created.}

\item{seed}{A single seed value to be passed to \code{set.seed()}.}

\item{...}{Additional params passed to \code{rsample::bootstraps()}.}
}
\description{
Fit and estimate variable importance from a workflow using many bootstrap resamples.
}
\examples{
\dontrun{
library(tidymodels)

# setup a workflow without fitting
wf <-
  workflow() \%>\%
  add_recipe(recipe(qsec ~ wt, data = mtcars)) \%>\%
  add_model(linear_reg())

# fit and estimate variable importance from 125 bootstrap resampled models
wf \%>\%
  vi_boots(training_data = mtcars, n_models = 125)
}
}
