% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{cov_exp_quad}
\alias{cov_exp_quad}
\title{Exponentiated Quadratic Covariance}
\usage{
cov_exp_quad(x, amplitude = 1, length_scale = 1, delta = 1e-09)
}
\arguments{
\item{x}{A vector containing positions}

\item{amplitude}{Numeric, controls the vertical scale of the covariance function}

\item{length_scale}{Numeric, controls the horizontal scale of the covariance function}

\item{delta}{A small offset added along the diagonal to ensure that the
function returns a positive-semidefinite matrix. Larger values allow for
increased variation at individual positions along the vector \code{x}.}
}
\value{
A \eqn{N \times N} symmetric, positive-semidefinite covariance matrix
}
\description{
Generate a \eqn{N \times N} covariance matrix, \eqn{\Sigma}, by
passing a distance vector \code{x} to the exponentiated quadratic kernel function:

\deqn{\Sigma_{ij} = \alpha^2 \exp \left( \frac{|x_i - x_j|^2}{-2 \ \rho} \right) + \delta_{ij} \ \sigma}

where \eqn{\alpha^2} is the \strong{amplitude}, \eqn{\rho} is the
\strong{length_scale}, and \eqn{\sigma} is a small offset along the diagonal to
allow for additional variation.
}
\examples{
x <- 1:12
cov_exp_quad(x)
}
