% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_extensions.R
\name{ggquicksave}
\alias{ggquicksave}
\title{Call ggsave with default parameters}
\usage{
ggquicksave(
  filename,
  plot = ggplot2::last_plot(),
  width = 9,
  height = 6,
  units = "in",
  dpi = 500,
  ...
)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{width, height, units}{Plot size in \code{units} ("in", "cm", "mm", or "px").}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320), "print" (300), or "screen" (72).}

\item{...}{Other arguments passed to \code{ggplot2::ggsave()}.}
}
\description{
Calls ggsave with some defaults that I happen to like. As of
\href{https://ggplot2.tidyverse.org/news/index.html#ggplot2-334}{ggplot2 3.3.4},
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} uses \href{https://ragg.r-lib.org/}{ragg} as the backend
device for saving.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point()

ggquicksave("mtcars.png")
}
}
