% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_transforms.R
\name{logit}
\alias{logit}
\title{Transform a value on a bound scale to an unbounded scale}
\usage{
logit(x, lower_bound = 0, upper_bound = 1, base = exp(1))
}
\arguments{
\item{x}{Bound value to be transformed}

\item{lower_bound}{Lower bound of x; defaults to 0.}

\item{upper_bound}{Upper bound of x; defaults to 1.}

\item{base}{Base used for log-transform, defaults to exp(1) = e}
}
\description{
Transform a value on a bound scale to an unbounded scale
}
\details{
Transform a value on a bound scale to an unbound scale, i.e.,
\verb{[-Inf, Inf]}. Values at the lower/upper bounds will return \code{-Inf}/\code{Inf}.
Uses the natural base, \code{e}, by default, but other bases can be passed to
the parameter \code{base}.
}
\examples{
\dontrun{
# unbound vector of values bound between [0, 1]
myvec <- seq(0, 10, 1)/10
logit(myvec)

# unbound vector of values between [100, 200]
myvec <- seq(100, 200, 10)
logit(myvec, lower_bound = 100, upper_bound = 200)
}
}
