% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse_extensions.R
\name{theme_rieke}
\alias{theme_rieke}
\title{Custom ggplot2 theme}
\usage{
theme_rieke(base_size = 14, base_family = "Roboto Slab")
}
\arguments{
\item{base_size}{base font size, given in pts}

\item{base_family}{base font family}
}
\description{
A custom ggplot2 theme based on my own personal preferences. This assumes the
user (me) is on a windows machine & has Roboto Slab already installed.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(title = "The title is left-aligned and allows for **bold** characters",
       subtitle = "You can also use any markdown formatting, including *italics*",
       caption = "Note the use of break<br>for line breaks")
}
}
